lithophane-app/requirements.txt
flask
numpy
cv2stl
skimage